+++ drivers/nut_usb.c
@@ -230,7 +230,7 @@ void nutusb_comm_fail(const char *fmt, ...)
 
 	if ((ret < 1) || (ret >= (int) sizeof(why)))
 		upslogx(LOG_WARNING, "usb_comm_fail: vsnprintf needed "
-			"more than %d bytes", sizeof(why));
+			"more than %d bytes", (int)sizeof(why));
 
 	upslogx(LOG_WARNING, "Communications with UPS lost: %s", why);
 }
