+++ drivers/safenet.c
@@ -89,7 +89,7 @@ static int safenet_command(const char *command)
 		return -1;
 	}
 
-	upsdebugx(3, "send: %.*s", strcspn(command, "\r"), command);
+	upsdebugx(3, "send: %.*s", (int)strcspn(command, "\r"), command);
 
 	/*
 	 * Read the reply from the UPS.
@@ -106,7 +106,7 @@ static int safenet_command(const char *command)
 		return -1;
 	}
 
-	upsdebugx(3, "read: %.*s", strcspn(reply, "\r"), reply);
+	upsdebugx(3, "read: %.*s", (int)strcspn(reply, "\r"), reply);
 
 	/*
 	 * We check if the reply looks like a valid status.
